Switch to using host.docker.internal
host.docker.internal is a host that points to the host machine.
Till now, we had to use a custom method to to add this host to
'/etc/hosts'. Docker 20.10 addded this natively to Linux and hence we
can get rid of this hack.

No clue if this will break things on Windows or MacOS.